Fix incorrect query results with spatial indexes [#1011453]
The bounding box computation for ellipses, polygons, and paths
appeared to be broken. As these bounding boxes are used by the
R-tree on-top-of GiST index implementation in pgsphere, the effect
was that some queries returned different results depending on
whether or not the optimizer chose to make use of an available
spatial index or not.

I have attached a patch which fixes these problems as best as I can tell.

http://pgfoundry.org/tracker/index.php?func=detail&aid=1011453&group_id=1000240&atid=914